Get path to snippets in a package

Get path to snippets in a package:

path_to_snippets_dir_of_pkg(package, ...)

path_to_snippets_files_of_pkg(package, ...)

Arguments

package
(character) Package name.

...
(character) Path to subdirectory with snippets of interest: {path to package}/inst/snippets/{path of subdirectory provided via ...}

Details

path_to_snippets_dir_of_pkg() gets path to directory with snippet files in a package. Defaults to {path to package}/inst/snippets. Returns empty string, if the directory does not exist.

path_to_snippets_files_of_pkg() gets paths to all files with snippets in a package. If the directory of interest does not exist, NULL is returned.

Examples

path_to_snippets_dir_of_pkg("snippets")

#> [1] "/Users/runner/work/_temp/Library/snippets/snippets"

path_to_snippets_files_of_pkg("snippets")

#> /Users/runner/work/_temp/Library/snippets/snippets/markdown.snippets
#> /Users/runner/work/_temp/Library/snippets/snippets/r.snippets